coll: clean up sched_auto method
The old algorithm selection functions, e.g. MPIR_Ibcast_intra_sched_auto
and MPIR_Ibcast_inter_sched_auto, have been replaced by csel selections.
Only a few places still need use sched auto to take care of sub
problems.

This commit moved the old sched auto code into to
src/mpi/coll/mpir_coll_sched_auto.c and removed its usage from csel
selections.

The general sched_auto method, that just calls intra_sched_auto or
inter_sched_auto, is removed.
