+++ src/mpi/coll/iallgather/iallgather_inter_sched_local_gather_remote_bcast.c
@@ -44,9 +44,9 @@ int MPIR_Iallgather_inter_sched_local_gather_remote_bcast(const void *sendbuf, M
     newcomm_ptr = comm_ptr->local_comm;
 
     if (sendcount != 0) {
-        mpi_errno = MPIR_Igather_sched_auto(sendbuf, sendcount, sendtype,
-                                            tmp_buf, sendcount * sendtype_sz, MPI_BYTE, 0,
-                                            newcomm_ptr, s);
+        mpi_errno = MPIR_Igather_intra_sched_auto(sendbuf, sendcount, sendtype,
+                                                  tmp_buf, sendcount * sendtype_sz, MPI_BYTE, 0,
+                                                  newcomm_ptr, s);
         MPIR_ERR_CHECK(mpi_errno);
         MPIR_SCHED_BARRIER(s);
     }
@@ -57,8 +57,8 @@ int MPIR_Iallgather_inter_sched_local_gather_remote_bcast(const void *sendbuf, M
         /* bcast to right */
         if (sendcount != 0) {
             root = (rank == 0) ? MPI_ROOT : MPI_PROC_NULL;
-            mpi_errno = MPIR_Ibcast_sched_auto(tmp_buf, sendcount * local_size * sendtype_sz,
-                                               MPI_BYTE, root, comm_ptr, s);
+            mpi_errno = MPIR_Ibcast_inter_sched_auto(tmp_buf, sendcount * local_size * sendtype_sz,
+                                                     MPI_BYTE, root, comm_ptr, s);
             MPIR_ERR_CHECK(mpi_errno);
         }
 
@@ -67,8 +67,8 @@ int MPIR_Iallgather_inter_sched_local_gather_remote_bcast(const void *sendbuf, M
         /* receive bcast from right */
         if (recvcount != 0) {
             root = 0;
-            mpi_errno = MPIR_Ibcast_sched_auto(recvbuf, recvcount * remote_size,
-                                               recvtype, root, comm_ptr, s);
+            mpi_errno = MPIR_Ibcast_inter_sched_auto(recvbuf, recvcount * remote_size,
+                                                     recvtype, root, comm_ptr, s);
             MPIR_ERR_CHECK(mpi_errno);
         }
         MPIR_SCHED_BARRIER(s);
@@ -76,8 +76,8 @@ int MPIR_Iallgather_inter_sched_local_gather_remote_bcast(const void *sendbuf, M
         /* receive bcast from left */
         if (recvcount != 0) {
             root = 0;
-            mpi_errno = MPIR_Ibcast_sched_auto(recvbuf, recvcount * remote_size,
-                                               recvtype, root, comm_ptr, s);
+            mpi_errno = MPIR_Ibcast_inter_sched_auto(recvbuf, recvcount * remote_size,
+                                                     recvtype, root, comm_ptr, s);
             MPIR_ERR_CHECK(mpi_errno);
         }
 
@@ -86,8 +86,8 @@ int MPIR_Iallgather_inter_sched_local_gather_remote_bcast(const void *sendbuf, M
         /* bcast to left */
         if (sendcount != 0) {
             root = (rank == 0) ? MPI_ROOT : MPI_PROC_NULL;
-            mpi_errno = MPIR_Ibcast_sched_auto(tmp_buf, sendcount * local_size * sendtype_sz,
-                                               MPI_BYTE, root, comm_ptr, s);
+            mpi_errno = MPIR_Ibcast_inter_sched_auto(tmp_buf, sendcount * local_size * sendtype_sz,
+                                                     MPI_BYTE, root, comm_ptr, s);
             MPIR_ERR_CHECK(mpi_errno);
         }
         MPIR_SCHED_BARRIER(s);
