+++ src/mpi/coll/ialltoallw/ialltoallw.c
@@ -59,13 +59,6 @@ int MPIR_Ialltoallw_allcomm_sched_auto(const void *sendbuf, const MPI_Aint * sen
                                                         comm_ptr, *sched_p);
             break;
 
-        case MPII_CSEL_CONTAINER_TYPE__ALGORITHM__MPIR_Ialltoallw_intra_sched_auto:
-            MPII_SCHED_CREATE_SCHED_P();
-            mpi_errno = MPIR_Ialltoallw_intra_sched_auto(sendbuf, sendcounts, sdispls, sendtypes,
-                                                         recvbuf, recvcounts, rdispls, recvtypes,
-                                                         comm_ptr, *sched_p);
-            break;
-
         case MPII_CSEL_CONTAINER_TYPE__ALGORITHM__MPIR_Ialltoallw_intra_sched_blocked:
             MPII_SCHED_CREATE_SCHED_P();
             mpi_errno = MPIR_Ialltoallw_intra_sched_blocked(sendbuf, sendcounts, sdispls, sendtypes,
@@ -78,13 +71,6 @@ int MPIR_Ialltoallw_allcomm_sched_auto(const void *sendbuf, const MPI_Aint * sen
             mpi_errno = MPIR_Ialltoallw_intra_sched_inplace(sendbuf, sendcounts, sdispls, sendtypes,
                                                             recvbuf, recvcounts, rdispls, recvtypes,
                                                             comm_ptr, *sched_p);
-            break;
-
-        case MPII_CSEL_CONTAINER_TYPE__ALGORITHM__MPIR_Ialltoallw_inter_sched_auto:
-            MPII_SCHED_CREATE_SCHED_P();
-            mpi_errno = MPIR_Ialltoallw_inter_sched_auto(sendbuf, sendcounts, sdispls, sendtypes,
-                                                         recvbuf, recvcounts, rdispls, recvtypes,
-                                                         comm_ptr, *sched_p);
             break;
 
         case MPII_CSEL_CONTAINER_TYPE__ALGORITHM__MPIR_Ialltoallw_inter_sched_pairwise_exchange:
@@ -105,62 +91,6 @@ int MPIR_Ialltoallw_allcomm_sched_auto(const void *sendbuf, const MPI_Aint * sen
     return mpi_errno;
   fn_fail:
     goto fn_exit;
-}
-
-int MPIR_Ialltoallw_intra_sched_auto(const void *sendbuf, const MPI_Aint sendcounts[],
-                                     const MPI_Aint sdispls[], const MPI_Datatype sendtypes[],
-                                     void *recvbuf, const MPI_Aint recvcounts[],
-                                     const MPI_Aint rdispls[], const MPI_Datatype recvtypes[],
-                                     MPIR_Comm * comm_ptr, MPIR_Sched_t s)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    if (sendbuf == MPI_IN_PLACE) {
-        mpi_errno = MPIR_Ialltoallw_intra_sched_inplace(sendbuf, sendcounts, sdispls,
-                                                        sendtypes, recvbuf, recvcounts,
-                                                        rdispls, recvtypes, comm_ptr, s);
-    } else {
-        mpi_errno = MPIR_Ialltoallw_intra_sched_blocked(sendbuf, sendcounts, sdispls,
-                                                        sendtypes, recvbuf, recvcounts,
-                                                        rdispls, recvtypes, comm_ptr, s);
-    }
-
-    return mpi_errno;
-}
-
-int MPIR_Ialltoallw_inter_sched_auto(const void *sendbuf, const MPI_Aint sendcounts[],
-                                     const MPI_Aint sdispls[], const MPI_Datatype sendtypes[],
-                                     void *recvbuf, const MPI_Aint recvcounts[],
-                                     const MPI_Aint rdispls[], const MPI_Datatype recvtypes[],
-                                     MPIR_Comm * comm_ptr, MPIR_Sched_t s)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    mpi_errno = MPIR_Ialltoallw_inter_sched_pairwise_exchange(sendbuf, sendcounts, sdispls,
-                                                              sendtypes, recvbuf, recvcounts,
-                                                              rdispls, recvtypes, comm_ptr, s);
-
-    return mpi_errno;
-}
-
-int MPIR_Ialltoallw_sched_auto(const void *sendbuf, const MPI_Aint sendcounts[],
-                               const MPI_Aint sdispls[], const MPI_Datatype sendtypes[],
-                               void *recvbuf, const MPI_Aint recvcounts[], const MPI_Aint rdispls[],
-                               const MPI_Datatype recvtypes[], MPIR_Comm * comm_ptr, MPIR_Sched_t s)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    if (comm_ptr->comm_kind == MPIR_COMM_KIND__INTRACOMM) {
-        mpi_errno = MPIR_Ialltoallw_intra_sched_auto(sendbuf, sendcounts, sdispls,
-                                                     sendtypes, recvbuf, recvcounts,
-                                                     rdispls, recvtypes, comm_ptr, s);
-    } else {
-        mpi_errno = MPIR_Ialltoallw_inter_sched_auto(sendbuf, sendcounts, sdispls,
-                                                     sendtypes, recvbuf, recvcounts,
-                                                     rdispls, recvtypes, comm_ptr, s);
-    }
-
-    return mpi_errno;
 }
 
 int MPIR_Ialltoallw_sched_impl(const void *sendbuf, const MPI_Aint sendcounts[],
@@ -217,14 +147,6 @@ int MPIR_Ialltoallw_sched_impl(const void *sendbuf, const MPI_Aint sendcounts[],
                                                                 *sched_p);
                 break;
 
-            case MPIR_CVAR_IALLTOALLW_INTRA_ALGORITHM_sched_auto:
-                MPII_SCHED_CREATE_SCHED_P();
-                mpi_errno = MPIR_Ialltoallw_intra_sched_auto(sendbuf, sendcounts, sdispls,
-                                                             sendtypes, recvbuf, recvcounts,
-                                                             rdispls, recvtypes, comm_ptr,
-                                                             *sched_p);
-                break;
-
             case MPIR_CVAR_IALLTOALLW_INTRA_ALGORITHM_auto:
                 mpi_errno =
                     MPIR_Ialltoallw_allcomm_sched_auto(sendbuf, sendcounts, sdispls, sendtypes,
@@ -249,14 +171,6 @@ int MPIR_Ialltoallw_sched_impl(const void *sendbuf, const MPI_Aint sendcounts[],
                                                                           comm_ptr, *sched_p);
                 break;
 
-            case MPIR_CVAR_IALLTOALLW_INTER_ALGORITHM_sched_auto:
-                MPII_SCHED_CREATE_SCHED_P();
-                mpi_errno = MPIR_Ialltoallw_inter_sched_auto(sendbuf, sendcounts, sdispls,
-                                                             sendtypes, recvbuf, recvcounts,
-                                                             rdispls, recvtypes, comm_ptr,
-                                                             *sched_p);
-                break;
-
             case MPIR_CVAR_IALLTOALLW_INTER_ALGORITHM_auto:
                 mpi_errno =
                     MPIR_Ialltoallw_allcomm_sched_auto(sendbuf, sendcounts, sdispls, sendtypes,
@@ -275,18 +189,9 @@ int MPIR_Ialltoallw_sched_impl(const void *sendbuf, const MPI_Aint sendcounts[],
     goto fn_exit;
 
   fallback:
-    if (comm_ptr->comm_kind == MPIR_COMM_KIND__INTRACOMM) {
-        MPII_SCHED_CREATE_SCHED_P();
-        mpi_errno = MPIR_Ialltoallw_intra_sched_auto(sendbuf, sendcounts, sdispls, sendtypes,
-                                                     recvbuf, recvcounts, rdispls, recvtypes,
-                                                     comm_ptr, *sched_p);
-    } else {
-        MPII_SCHED_CREATE_SCHED_P();
-        mpi_errno = MPIR_Ialltoallw_inter_sched_auto(sendbuf, sendcounts, sdispls, sendtypes,
-                                                     recvbuf, recvcounts, rdispls, recvtypes,
-                                                     comm_ptr, *sched_p);
-    }
-
+    mpi_errno = MPIR_Ialltoallw_allcomm_sched_auto(sendbuf, sendcounts, sdispls, sendtypes,
+                                                   recvbuf, recvcounts, rdispls, recvtypes,
+                                                   comm_ptr, is_persistent, sched_p, sched_type_p);
   fn_exit:
     return mpi_errno;
   fn_fail:
