+++ src/mpi/coll/ibarrier/ibarrier.c
@@ -30,11 +30,6 @@ int MPIR_Ibarrier_allcomm_sched_auto(MPIR_Comm * comm_ptr, bool is_persistent, v
 
     switch (cnt->id) {
         /* *INDENT-OFF* */
-        case MPII_CSEL_CONTAINER_TYPE__ALGORITHM__MPIR_Ibarrier_intra_sched_auto:
-            MPII_SCHED_CREATE_SCHED_P();
-            mpi_errno = MPIR_Ibarrier_intra_sched_auto(comm_ptr, *sched_p);
-            break;
-
         case MPII_CSEL_CONTAINER_TYPE__ALGORITHM__MPIR_Ibarrier_intra_sched_recursive_doubling:
             MPII_SCHED_CREATE_SCHED_P();
             mpi_errno = MPIR_Ibarrier_intra_sched_recursive_doubling(comm_ptr, *sched_p);
@@ -48,11 +43,6 @@ int MPIR_Ibarrier_allcomm_sched_auto(MPIR_Comm * comm_ptr, bool is_persistent, v
                                                         MPIR_IALLREDUCE_RECEXCH_TYPE_MULTIPLE_BUFFER,
                                                         cnt->u.ibarrier.intra_gentran_recexch.k,
                                                         *sched_p);
-            break;
-
-        case MPII_CSEL_CONTAINER_TYPE__ALGORITHM__MPIR_Ibarrier_inter_sched_auto:
-            MPII_SCHED_CREATE_SCHED_P();
-            mpi_errno = MPIR_Ibarrier_inter_sched_auto(comm_ptr, *sched_p);
             break;
 
         case MPII_CSEL_CONTAINER_TYPE__ALGORITHM__MPIR_Ibarrier_inter_sched_bcast:
@@ -69,39 +59,6 @@ int MPIR_Ibarrier_allcomm_sched_auto(MPIR_Comm * comm_ptr, bool is_persistent, v
     return mpi_errno;
   fn_fail:
     goto fn_exit;
-}
-
-int MPIR_Ibarrier_intra_sched_auto(MPIR_Comm * comm_ptr, MPIR_Sched_t s)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    mpi_errno = MPIR_Ibarrier_intra_sched_recursive_doubling(comm_ptr, s);
-
-    return mpi_errno;
-}
-
-/* It will choose between several different algorithms based on the given
- * parameters. */
-int MPIR_Ibarrier_inter_sched_auto(MPIR_Comm * comm_ptr, MPIR_Sched_t s)
-{
-    int mpi_errno;
-
-    mpi_errno = MPIR_Ibarrier_inter_sched_bcast(comm_ptr, s);
-
-    return mpi_errno;
-}
-
-int MPIR_Ibarrier_sched_auto(MPIR_Comm * comm_ptr, MPIR_Sched_t s)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    if (comm_ptr->comm_kind == MPIR_COMM_KIND__INTRACOMM) {
-        mpi_errno = MPIR_Ibarrier_intra_sched_auto(comm_ptr, s);
-    } else {
-        mpi_errno = MPIR_Ibarrier_inter_sched_auto(comm_ptr, s);
-    }
-
-    return mpi_errno;
 }
 
 int MPIR_Ibarrier_sched_impl(MPIR_Comm * comm_ptr, bool is_persistent, void **sched_p,
@@ -134,11 +91,6 @@ int MPIR_Ibarrier_sched_impl(MPIR_Comm * comm_ptr, bool is_persistent, void **sc
                 mpi_errno = MPIR_Ibarrier_intra_sched_recursive_doubling(comm_ptr, *sched_p);
                 break;
 
-            case MPIR_CVAR_IBARRIER_INTRA_ALGORITHM_sched_auto:
-                MPII_SCHED_CREATE_SCHED_P();
-                mpi_errno = MPIR_Ibarrier_intra_sched_auto(comm_ptr, *sched_p);
-                break;
-
             case MPIR_CVAR_IBARRIER_INTRA_ALGORITHM_auto:
                 mpi_errno = MPIR_Ibarrier_allcomm_sched_auto(comm_ptr, is_persistent, sched_p,
                                                              sched_type_p);
@@ -154,11 +106,6 @@ int MPIR_Ibarrier_sched_impl(MPIR_Comm * comm_ptr, bool is_persistent, void **sc
             case MPIR_CVAR_IBARRIER_INTER_ALGORITHM_sched_bcast:
                 MPII_SCHED_CREATE_SCHED_P();
                 mpi_errno = MPIR_Ibarrier_inter_sched_bcast(comm_ptr, *sched_p);
-                break;
-
-            case MPIR_CVAR_IBARRIER_INTER_ALGORITHM_sched_auto:
-                MPII_SCHED_CREATE_SCHED_P();
-                mpi_errno = MPIR_Ibarrier_inter_sched_auto(comm_ptr, *sched_p);
                 break;
 
             case MPIR_CVAR_IBARRIER_INTER_ALGORITHM_auto:
