+++ src/mpi/coll/ibcast/ibcast_inter_sched_flat.c
@@ -38,7 +38,8 @@ int MPIR_Ibcast_inter_sched_flat(void *buffer, MPI_Aint count, MPI_Datatype data
 
         /* now do the usual broadcast on this intracommunicator
          * with rank 0 as root. */
-        mpi_errno = MPIR_Ibcast_sched_auto(buffer, count, datatype, root, comm_ptr->local_comm, s);
+        mpi_errno =
+            MPIR_Ibcast_intra_sched_auto(buffer, count, datatype, root, comm_ptr->local_comm, s);
         MPIR_ERR_CHECK(mpi_errno);
     }
