+++ src/mpi/coll/ibcast/ibcast_intra_sched_scatter_recursive_doubling_allgather.c
@@ -30,8 +30,8 @@
  */
 /* It would be nice to just call:
  * ----8<----
- * MPIR_Iscatter_sched_auto(...);
- * MPIR_Iallgather_sched_auto(...);
+ * MPIR_Iscatter_intra_sched_auto(...);
+ * MPIR_Iallgather_intra_sched_auto(...);
  * ----8<----
  *
  * But that results in inefficient additional memory allocation and copies
