+++ src/mpi/coll/ibcast/ibcast_intra_sched_smp.c
@@ -65,9 +65,9 @@ int MPIR_Ibcast_intra_sched_smp(void *buffer, MPI_Aint count, MPI_Datatype datat
 
     /* perform the internode broadcast */
     if (comm_ptr->node_roots_comm != NULL) {
-        mpi_errno = MPIR_Ibcast_sched_auto(buffer, count, datatype,
-                                           MPIR_Get_internode_rank(comm_ptr, root),
-                                           comm_ptr->node_roots_comm, s);
+        mpi_errno = MPIR_Ibcast_intra_sched_auto(buffer, count, datatype,
+                                                 MPIR_Get_internode_rank(comm_ptr, root),
+                                                 comm_ptr->node_roots_comm, s);
         MPIR_ERR_CHECK(mpi_errno);
 
         /* don't allow the local ops for the intranode phase to start until this has completed */
@@ -75,7 +75,8 @@ int MPIR_Ibcast_intra_sched_smp(void *buffer, MPI_Aint count, MPI_Datatype datat
     }
     /* perform the intranode broadcast on all except for the root's node */
     if (comm_ptr->node_comm != NULL) {
-        mpi_errno = MPIR_Ibcast_sched_auto(buffer, count, datatype, 0, comm_ptr->node_comm, s);
+        mpi_errno =
+            MPIR_Ibcast_intra_sched_auto(buffer, count, datatype, 0, comm_ptr->node_comm, s);
         MPIR_ERR_CHECK(mpi_errno);
     }
