+++ src/mpi/coll/igather/igather_inter_sched_short.c
@@ -46,7 +46,7 @@ int MPIR_Igather_inter_sched_short(const void *sendbuf, MPI_Aint sendcount, MPI_
             tmp_buf = MPIR_Sched_alloc_state(s, sendcount * local_size * sendtype_sz);
             MPIR_ERR_CHKANDJUMP(!tmp_buf, mpi_errno, MPI_ERR_OTHER, "**nomem");
         } else {
-            /* silience -Wmaybe-uninitialized due to MPIR_Igather_sched_auto by non-zero ranks */
+            /* silience -Wmaybe-uninitialized due to MPIR_Igather_intra_sched_auto by non-zero ranks */
             sendtype_sz = 0;
         }
 
@@ -59,9 +59,9 @@ int MPIR_Igather_inter_sched_short(const void *sendbuf, MPI_Aint sendcount, MPI_
         newcomm_ptr = comm_ptr->local_comm;
 
         /* now do the a local gather on this intracommunicator */
-        mpi_errno = MPIR_Igather_sched_auto(sendbuf, sendcount, sendtype,
-                                            tmp_buf, sendcount * sendtype_sz, MPI_BYTE, 0,
-                                            newcomm_ptr, s);
+        mpi_errno = MPIR_Igather_intra_sched_auto(sendbuf, sendcount, sendtype,
+                                                  tmp_buf, sendcount * sendtype_sz, MPI_BYTE, 0,
+                                                  newcomm_ptr, s);
         MPIR_ERR_CHECK(mpi_errno);
 
         if (rank == 0) {
