+++ src/mpi/coll/ineighbor_alltoallv/ineighbor_alltoallv.c
@@ -49,22 +49,6 @@ int MPIR_Ineighbor_alltoallv_allcomm_sched_auto(const void *sendbuf, const MPI_A
                                                                   *sched_p);
             break;
 
-        case MPII_CSEL_CONTAINER_TYPE__ALGORITHM__MPIR_Ineighbor_alltoallv_intra_sched_auto:
-            MPII_SCHED_CREATE_SCHED_P();
-            mpi_errno =
-                MPIR_Ineighbor_alltoallv_intra_sched_auto(sendbuf, sendcounts, sdispls, sendtype,
-                                                          recvbuf, recvcounts, rdispls, recvtype,
-                                                          comm_ptr, *sched_p);
-            break;
-
-        case MPII_CSEL_CONTAINER_TYPE__ALGORITHM__MPIR_Ineighbor_alltoallv_inter_sched_auto:
-            MPII_SCHED_CREATE_SCHED_P();
-            mpi_errno =
-                MPIR_Ineighbor_alltoallv_inter_sched_auto(sendbuf, sendcounts, sdispls, sendtype,
-                                                          recvbuf, recvcounts, rdispls, recvtype,
-                                                          comm_ptr, *sched_p);
-            break;
-
         case MPII_CSEL_CONTAINER_TYPE__ALGORITHM__MPIR_Ineighbor_alltoallv_allcomm_sched_linear:
             MPII_SCHED_CREATE_SCHED_P();
             mpi_errno =
@@ -83,71 +67,6 @@ int MPIR_Ineighbor_alltoallv_allcomm_sched_auto(const void *sendbuf, const MPI_A
     return mpi_errno;
   fn_fail:
     goto fn_exit;
-}
-
-int MPIR_Ineighbor_alltoallv_intra_sched_auto(const void *sendbuf, const MPI_Aint sendcounts[],
-                                              const MPI_Aint sdispls[], MPI_Datatype sendtype,
-                                              void *recvbuf, const MPI_Aint recvcounts[],
-                                              const MPI_Aint rdispls[], MPI_Datatype recvtype,
-                                              MPIR_Comm * comm_ptr, MPIR_Sched_t s)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    mpi_errno =
-        MPIR_Ineighbor_alltoallv_allcomm_sched_linear(sendbuf, sendcounts, sdispls, sendtype,
-                                                      recvbuf, recvcounts, rdispls, recvtype,
-                                                      comm_ptr, s);
-    MPIR_ERR_CHECK(mpi_errno);
-
-  fn_exit:
-    return mpi_errno;
-
-  fn_fail:
-    goto fn_exit;
-}
-
-int MPIR_Ineighbor_alltoallv_inter_sched_auto(const void *sendbuf, const MPI_Aint sendcounts[],
-                                              const MPI_Aint sdispls[], MPI_Datatype sendtype,
-                                              void *recvbuf, const MPI_Aint recvcounts[],
-                                              const MPI_Aint rdispls[], MPI_Datatype recvtype,
-                                              MPIR_Comm * comm_ptr, MPIR_Sched_t s)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    mpi_errno =
-        MPIR_Ineighbor_alltoallv_allcomm_sched_linear(sendbuf, sendcounts, sdispls, sendtype,
-                                                      recvbuf, recvcounts, rdispls, recvtype,
-                                                      comm_ptr, s);
-    MPIR_ERR_CHECK(mpi_errno);
-
-  fn_exit:
-    return mpi_errno;
-
-  fn_fail:
-    goto fn_exit;
-}
-
-int MPIR_Ineighbor_alltoallv_sched_auto(const void *sendbuf, const MPI_Aint sendcounts[],
-                                        const MPI_Aint sdispls[], MPI_Datatype sendtype,
-                                        void *recvbuf, const MPI_Aint recvcounts[],
-                                        const MPI_Aint rdispls[], MPI_Datatype recvtype,
-                                        MPIR_Comm * comm_ptr, MPIR_Sched_t s)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    if (comm_ptr->comm_kind == MPIR_COMM_KIND__INTRACOMM) {
-        mpi_errno =
-            MPIR_Ineighbor_alltoallv_intra_sched_auto(sendbuf, sendcounts, sdispls,
-                                                      sendtype, recvbuf, recvcounts,
-                                                      rdispls, recvtype, comm_ptr, s);
-    } else {
-        mpi_errno =
-            MPIR_Ineighbor_alltoallv_inter_sched_auto(sendbuf, sendcounts, sdispls,
-                                                      sendtype, recvbuf, recvcounts,
-                                                      rdispls, recvtype, comm_ptr, s);
-    }
-
-    return mpi_errno;
 }
 
 int MPIR_Ineighbor_alltoallv_sched_impl(const void *sendbuf, const MPI_Aint sendcounts[],
@@ -185,14 +104,6 @@ int MPIR_Ineighbor_alltoallv_sched_impl(const void *sendbuf, const MPI_Aint send
                                                                   *sched_p);
                 break;
 
-            case MPIR_CVAR_INEIGHBOR_ALLTOALLV_INTRA_ALGORITHM_sched_auto:
-                MPII_SCHED_CREATE_SCHED_P();
-                mpi_errno = MPIR_Ineighbor_alltoallv_intra_sched_auto(sendbuf, sendcounts, sdispls,
-                                                                      sendtype, recvbuf, recvcounts,
-                                                                      rdispls, recvtype, comm_ptr,
-                                                                      *sched_p);
-                break;
-
             case MPIR_CVAR_INEIGHBOR_ALLTOALLV_INTRA_ALGORITHM_auto:
                 mpi_errno =
                     MPIR_Ineighbor_alltoallv_allcomm_sched_auto(sendbuf, sendcounts, sdispls,
@@ -225,15 +136,6 @@ int MPIR_Ineighbor_alltoallv_sched_impl(const void *sendbuf, const MPI_Aint send
                                                                   sendtype, recvbuf, recvcounts,
                                                                   rdispls, recvtype, comm_ptr,
                                                                   *sched_p);
-                break;
-
-            case MPIR_CVAR_INEIGHBOR_ALLTOALLV_INTER_ALGORITHM_sched_auto:
-                MPII_SCHED_CREATE_SCHED_P();
-                mpi_errno =
-                    MPIR_Ineighbor_alltoallv_inter_sched_auto(sendbuf, sendcounts, sdispls,
-                                                              sendtype, recvbuf, recvcounts,
-                                                              rdispls, recvtype, comm_ptr,
-                                                              *sched_p);
                 break;
 
             case MPIR_CVAR_INEIGHBOR_ALLTOALLV_INTER_ALGORITHM_auto:
