+++ src/mpi/coll/ireduce/ireduce_intra_sched_smp.c
@@ -44,7 +44,7 @@ int MPIR_Ireduce_intra_sched_smp(const void *sendbuf, void *recvbuf, MPI_Aint co
 
     /* do the intranode reduce on all nodes other than the root's node */
     if (nc != NULL && MPIR_Get_intranode_rank(comm_ptr, root) == -1) {
-        mpi_errno = MPIR_Ireduce_sched_auto(sendbuf, tmp_buf, count, datatype, op, 0, nc, s);
+        mpi_errno = MPIR_Ireduce_intra_sched_auto(sendbuf, tmp_buf, count, datatype, op, 0, nc, s);
         MPIR_ERR_CHECK(mpi_errno);
         MPIR_SCHED_BARRIER(s);
     }
@@ -55,9 +55,9 @@ int MPIR_Ireduce_intra_sched_smp(const void *sendbuf, void *recvbuf, MPI_Aint co
             /* I am not on root's node.  Use tmp_buf if we
              * participated in the first reduce, otherwise use sendbuf */
             const void *buf = (nc == NULL ? sendbuf : tmp_buf);
-            mpi_errno = MPIR_Ireduce_sched_auto(buf, NULL, count, datatype,
-                                                op, MPIR_Get_internode_rank(comm_ptr, root), nrc,
-                                                s);
+            mpi_errno = MPIR_Ireduce_intra_sched_auto(buf, NULL, count, datatype,
+                                                      op, MPIR_Get_internode_rank(comm_ptr, root),
+                                                      nrc, s);
             MPIR_ERR_CHECK(mpi_errno);
             MPIR_SCHED_BARRIER(s);
         } else {        /* I am on root's node. I have not participated in the earlier reduce. */
@@ -65,9 +65,10 @@ int MPIR_Ireduce_intra_sched_smp(const void *sendbuf, void *recvbuf, MPI_Aint co
                 /* I am not the root though. I don't have a valid recvbuf.
                  * Use tmp_buf as recvbuf. */
 
-                mpi_errno = MPIR_Ireduce_sched_auto(sendbuf, tmp_buf, count, datatype,
-                                                    op, MPIR_Get_internode_rank(comm_ptr, root),
-                                                    nrc, s);
+                mpi_errno = MPIR_Ireduce_intra_sched_auto(sendbuf, tmp_buf, count, datatype,
+                                                          op, MPIR_Get_internode_rank(comm_ptr,
+                                                                                      root), nrc,
+                                                          s);
                 MPIR_ERR_CHECK(mpi_errno);
                 MPIR_SCHED_BARRIER(s);
 
@@ -76,9 +77,10 @@ int MPIR_Ireduce_intra_sched_smp(const void *sendbuf, void *recvbuf, MPI_Aint co
             } else {
                 /* I am the root. in_place is automatically handled. */
 
-                mpi_errno = MPIR_Ireduce_sched_auto(sendbuf, recvbuf, count, datatype,
-                                                    op, MPIR_Get_internode_rank(comm_ptr, root),
-                                                    nrc, s);
+                mpi_errno = MPIR_Ireduce_intra_sched_auto(sendbuf, recvbuf, count, datatype,
+                                                          op, MPIR_Get_internode_rank(comm_ptr,
+                                                                                      root), nrc,
+                                                          s);
                 MPIR_ERR_CHECK(mpi_errno);
                 MPIR_SCHED_BARRIER(s);
 
@@ -90,8 +92,9 @@ int MPIR_Ireduce_intra_sched_smp(const void *sendbuf, void *recvbuf, MPI_Aint co
 
     /* do the intranode reduce on the root's node */
     if (nc != NULL && MPIR_Get_intranode_rank(comm_ptr, root) != -1) {
-        mpi_errno = MPIR_Ireduce_sched_auto(sendbuf, recvbuf, count, datatype,
-                                            op, MPIR_Get_intranode_rank(comm_ptr, root), nc, s);
+        mpi_errno = MPIR_Ireduce_intra_sched_auto(sendbuf, recvbuf, count, datatype,
+                                                  op, MPIR_Get_intranode_rank(comm_ptr, root), nc,
+                                                  s);
         MPIR_ERR_CHECK(mpi_errno);
         MPIR_SCHED_BARRIER(s);
     }
