+++ src/mpi/coll/ireduce_scatter/ireduce_scatter.c
@@ -36,12 +36,6 @@ int MPIR_Ireduce_scatter_allcomm_sched_auto(const void *sendbuf, void *recvbuf,
 
     switch (cnt->id) {
         /* *INDENT-OFF* */
-        case MPII_CSEL_CONTAINER_TYPE__ALGORITHM__MPIR_Ireduce_scatter_intra_sched_auto:
-            MPII_SCHED_CREATE_SCHED_P();
-            mpi_errno = MPIR_Ireduce_scatter_intra_sched_auto(sendbuf, recvbuf, recvcounts,
-                                                              datatype, op, comm_ptr, *sched_p);
-            break;
-
         case MPII_CSEL_CONTAINER_TYPE__ALGORITHM__MPIR_Ireduce_scatter_intra_sched_noncommutative:
             MPII_SCHED_CREATE_SCHED_P();
             mpi_errno = MPIR_Ireduce_scatter_intra_sched_noncommutative(sendbuf, recvbuf,
@@ -82,12 +76,6 @@ int MPIR_Ireduce_scatter_allcomm_sched_auto(const void *sendbuf, void *recvbuf,
                                                              *sched_p);
             break;
 
-        case MPII_CSEL_CONTAINER_TYPE__ALGORITHM__MPIR_Ireduce_scatter_inter_sched_auto:
-            MPII_SCHED_CREATE_SCHED_P();
-            mpi_errno = MPIR_Ireduce_scatter_inter_sched_auto(sendbuf, recvbuf, recvcounts,
-                                                              datatype, op, comm_ptr, *sched_p);
-            break;
-
         case MPII_CSEL_CONTAINER_TYPE__ALGORITHM__MPIR_Ireduce_scatter_inter_sched_remote_reduce_local_scatterv:
             MPII_SCHED_CREATE_SCHED_P();
             mpi_errno =
@@ -106,101 +94,6 @@ int MPIR_Ireduce_scatter_allcomm_sched_auto(const void *sendbuf, void *recvbuf,
     return mpi_errno;
   fn_fail:
     goto fn_exit;
-}
-
-int MPIR_Ireduce_scatter_intra_sched_auto(const void *sendbuf, void *recvbuf,
-                                          const MPI_Aint recvcounts[], MPI_Datatype datatype,
-                                          MPI_Op op, MPIR_Comm * comm_ptr, MPIR_Sched_t s)
-{
-    int mpi_errno = MPI_SUCCESS;
-    int i;
-    int is_commutative;
-    int total_count, type_size, nbytes;
-    int comm_size;
-
-    is_commutative = MPIR_Op_is_commutative(op);
-
-    comm_size = comm_ptr->local_size;
-    total_count = 0;
-    for (i = 0; i < comm_size; i++) {
-        total_count += recvcounts[i];
-    }
-    if (total_count == 0) {
-        goto fn_exit;
-    }
-    MPIR_Datatype_get_size_macro(datatype, type_size);
-    nbytes = total_count * type_size;
-
-    /* select an appropriate algorithm based on commutivity and message size */
-    if (is_commutative && (nbytes < MPIR_CVAR_REDUCE_SCATTER_COMMUTATIVE_LONG_MSG_SIZE)) {
-        mpi_errno =
-            MPIR_Ireduce_scatter_intra_sched_recursive_halving(sendbuf, recvbuf, recvcounts,
-                                                               datatype, op, comm_ptr, s);
-        MPIR_ERR_CHECK(mpi_errno);
-    } else if (is_commutative && (nbytes >= MPIR_CVAR_REDUCE_SCATTER_COMMUTATIVE_LONG_MSG_SIZE)) {
-        mpi_errno =
-            MPIR_Ireduce_scatter_intra_sched_pairwise(sendbuf, recvbuf, recvcounts, datatype, op,
-                                                      comm_ptr, s);
-        MPIR_ERR_CHECK(mpi_errno);
-    } else {    /* (!is_commutative) */
-
-        int is_block_regular = TRUE;
-        for (i = 0; i < (comm_size - 1); ++i) {
-            if (recvcounts[i] != recvcounts[i + 1]) {
-                is_block_regular = FALSE;
-                break;
-            }
-        }
-
-        if (MPL_is_pof2(comm_size, NULL) && is_block_regular) {
-            /* noncommutative, pof2 size, and block regular */
-            mpi_errno =
-                MPIR_Ireduce_scatter_intra_sched_noncommutative(sendbuf, recvbuf, recvcounts,
-                                                                datatype, op, comm_ptr, s);
-            MPIR_ERR_CHECK(mpi_errno);
-        } else {
-            /* noncommutative and (non-pof2 or block irregular), use recursive doubling. */
-            mpi_errno =
-                MPIR_Ireduce_scatter_intra_sched_recursive_doubling(sendbuf, recvbuf, recvcounts,
-                                                                    datatype, op, comm_ptr, s);
-            MPIR_ERR_CHECK(mpi_errno);
-        }
-    }
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-int MPIR_Ireduce_scatter_inter_sched_auto(const void *sendbuf, void *recvbuf,
-                                          const MPI_Aint recvcounts[], MPI_Datatype datatype,
-                                          MPI_Op op, MPIR_Comm * comm_ptr, MPIR_Sched_t s)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    mpi_errno =
-        MPIR_Ireduce_scatter_inter_sched_remote_reduce_local_scatterv(sendbuf, recvbuf, recvcounts,
-                                                                      datatype, op, comm_ptr, s);
-
-    return mpi_errno;
-}
-
-int MPIR_Ireduce_scatter_sched_auto(const void *sendbuf, void *recvbuf, const MPI_Aint recvcounts[],
-                                    MPI_Datatype datatype, MPI_Op op, MPIR_Comm * comm_ptr,
-                                    MPIR_Sched_t s)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    if (comm_ptr->comm_kind == MPIR_COMM_KIND__INTRACOMM) {
-        mpi_errno = MPIR_Ireduce_scatter_intra_sched_auto(sendbuf, recvbuf,
-                                                          recvcounts, datatype, op, comm_ptr, s);
-    } else {
-        mpi_errno = MPIR_Ireduce_scatter_inter_sched_auto(sendbuf, recvbuf, recvcounts,
-                                                          datatype, op, comm_ptr, s);
-    }
-
-    return mpi_errno;
 }
 
 int MPIR_Ireduce_scatter_sched_impl(const void *sendbuf, void *recvbuf, const MPI_Aint recvcounts[],
@@ -261,12 +154,6 @@ int MPIR_Ireduce_scatter_sched_impl(const void *sendbuf, void *recvbuf, const MP
                                                                                 comm_ptr, *sched_p);
                 break;
 
-            case MPIR_CVAR_IREDUCE_SCATTER_INTRA_ALGORITHM_sched_auto:
-                MPII_SCHED_CREATE_SCHED_P();
-                mpi_errno = MPIR_Ireduce_scatter_intra_sched_auto(sendbuf, recvbuf, recvcounts,
-                                                                  datatype, op, comm_ptr, *sched_p);
-                break;
-
             case MPIR_CVAR_IREDUCE_SCATTER_INTRA_ALGORITHM_auto:
                 mpi_errno =
                     MPIR_Ireduce_scatter_allcomm_sched_auto(sendbuf, recvbuf, recvcounts, datatype,
@@ -291,12 +178,6 @@ int MPIR_Ireduce_scatter_sched_impl(const void *sendbuf, void *recvbuf, const MP
                                                                                   *sched_p);
                 break;
 
-            case MPIR_CVAR_IREDUCE_SCATTER_INTER_ALGORITHM_sched_auto:
-                MPII_SCHED_CREATE_SCHED_P();
-                mpi_errno = MPIR_Ireduce_scatter_inter_sched_auto(sendbuf, recvbuf, recvcounts,
-                                                                  datatype, op, comm_ptr, *sched_p);
-                break;
-
             case MPIR_CVAR_IREDUCE_SCATTER_INTER_ALGORITHM_auto:
                 mpi_errno =
                     MPIR_Ireduce_scatter_allcomm_sched_auto(sendbuf, recvbuf, recvcounts, datatype,
@@ -314,16 +195,9 @@ int MPIR_Ireduce_scatter_sched_impl(const void *sendbuf, void *recvbuf, const MP
     goto fn_exit;
 
   fallback:
-    if (comm_ptr->comm_kind == MPIR_COMM_KIND__INTRACOMM) {
-        MPII_SCHED_CREATE_SCHED_P();
-        mpi_errno = MPIR_Ireduce_scatter_intra_sched_auto(sendbuf, recvbuf, recvcounts, datatype,
-                                                          op, comm_ptr, *sched_p);
-    } else {
-        MPII_SCHED_CREATE_SCHED_P();
-        mpi_errno = MPIR_Ireduce_scatter_inter_sched_auto(sendbuf, recvbuf, recvcounts, datatype,
-                                                          op, comm_ptr, *sched_p);
-    }
-
+    mpi_errno = MPIR_Ireduce_scatter_allcomm_sched_auto(sendbuf, recvbuf, recvcounts, datatype,
+                                                        op, comm_ptr, is_persistent, sched_p,
+                                                        sched_type_p);
   fn_exit:
     return mpi_errno;
   fn_fail:
