+++ src/mpi/coll/ireduce_scatter/ireduce_scatter_inter_sched_remote_reduce_local_scatterv.c
@@ -61,8 +61,8 @@ int MPIR_Ireduce_scatter_inter_sched_remote_reduce_local_scatterv(const void *se
     if (comm_ptr->is_low_group) {
         /* reduce from right group to rank 0 */
         root = (rank == 0) ? MPI_ROOT : MPI_PROC_NULL;
-        mpi_errno = MPIR_Ireduce_sched_auto(sendbuf, tmp_buf, total_count,
-                                            datatype, op, root, comm_ptr, s);
+        mpi_errno = MPIR_Ireduce_inter_sched_auto(sendbuf, tmp_buf, total_count,
+                                                  datatype, op, root, comm_ptr, s);
         MPIR_ERR_CHECK(mpi_errno);
 
         /* sched barrier intentionally omitted here to allow both reductions to
@@ -70,14 +70,14 @@ int MPIR_Ireduce_scatter_inter_sched_remote_reduce_local_scatterv(const void *se
 
         /* reduce to rank 0 of right group */
         root = 0;
-        mpi_errno = MPIR_Ireduce_sched_auto(sendbuf, tmp_buf, total_count,
-                                            datatype, op, root, comm_ptr, s);
+        mpi_errno = MPIR_Ireduce_inter_sched_auto(sendbuf, tmp_buf, total_count,
+                                                  datatype, op, root, comm_ptr, s);
         MPIR_ERR_CHECK(mpi_errno);
     } else {
         /* reduce to rank 0 of right group */
         root = 0;
-        mpi_errno = MPIR_Ireduce_sched_auto(sendbuf, tmp_buf, total_count,
-                                            datatype, op, root, comm_ptr, s);
+        mpi_errno = MPIR_Ireduce_inter_sched_auto(sendbuf, tmp_buf, total_count,
+                                                  datatype, op, root, comm_ptr, s);
         MPIR_ERR_CHECK(mpi_errno);
 
         /* sched barrier intentionally omitted here to allow both reductions to
@@ -85,8 +85,8 @@ int MPIR_Ireduce_scatter_inter_sched_remote_reduce_local_scatterv(const void *se
 
         /* reduce from right group to rank 0 */
         root = (rank == 0) ? MPI_ROOT : MPI_PROC_NULL;
-        mpi_errno = MPIR_Ireduce_sched_auto(sendbuf, tmp_buf, total_count,
-                                            datatype, op, root, comm_ptr, s);
+        mpi_errno = MPIR_Ireduce_inter_sched_auto(sendbuf, tmp_buf, total_count,
+                                                  datatype, op, root, comm_ptr, s);
         MPIR_ERR_CHECK(mpi_errno);
     }
     MPIR_SCHED_BARRIER(s);
@@ -99,8 +99,9 @@ int MPIR_Ireduce_scatter_inter_sched_remote_reduce_local_scatterv(const void *se
 
     newcomm_ptr = comm_ptr->local_comm;
 
-    mpi_errno = MPIR_Iscatterv_sched_auto(tmp_buf, recvcounts, disps, datatype,
-                                          recvbuf, recvcounts[rank], datatype, 0, newcomm_ptr, s);
+    mpi_errno = MPIR_Iscatterv_intra_sched_auto(tmp_buf, recvcounts, disps, datatype,
+                                                recvbuf, recvcounts[rank], datatype, 0, newcomm_ptr,
+                                                s);
     MPIR_ERR_CHECK(mpi_errno);
 
   fn_exit:
