+++ src/mpi/coll/ireduce_scatter_block/ireduce_scatter_block_inter_sched_remote_reduce_local_scatterv.c
@@ -49,8 +49,8 @@ int MPIR_Ireduce_scatter_block_inter_sched_remote_reduce_local_scatterv(const vo
     if (comm_ptr->is_low_group) {
         /* reduce from right group to rank 0 */
         root = (rank == 0) ? MPI_ROOT : MPI_PROC_NULL;
-        mpi_errno = MPIR_Ireduce_sched_auto(sendbuf, tmp_buf, total_count,
-                                            datatype, op, root, comm_ptr, s);
+        mpi_errno = MPIR_Ireduce_inter_sched_auto(sendbuf, tmp_buf, total_count,
+                                                  datatype, op, root, comm_ptr, s);
         MPIR_ERR_CHECK(mpi_errno);
 
         /* sched barrier intentionally omitted here to allow both reductions to
@@ -58,14 +58,14 @@ int MPIR_Ireduce_scatter_block_inter_sched_remote_reduce_local_scatterv(const vo
 
         /* reduce to rank 0 of right group */
         root = 0;
-        mpi_errno = MPIR_Ireduce_sched_auto(sendbuf, tmp_buf, total_count,
-                                            datatype, op, root, comm_ptr, s);
+        mpi_errno = MPIR_Ireduce_inter_sched_auto(sendbuf, tmp_buf, total_count,
+                                                  datatype, op, root, comm_ptr, s);
         MPIR_ERR_CHECK(mpi_errno);
     } else {
         /* reduce to rank 0 of right group */
         root = 0;
-        mpi_errno = MPIR_Ireduce_sched_auto(sendbuf, tmp_buf, total_count,
-                                            datatype, op, root, comm_ptr, s);
+        mpi_errno = MPIR_Ireduce_inter_sched_auto(sendbuf, tmp_buf, total_count,
+                                                  datatype, op, root, comm_ptr, s);
         MPIR_ERR_CHECK(mpi_errno);
 
         /* sched barrier intentionally omitted here to allow both reductions to
@@ -73,8 +73,8 @@ int MPIR_Ireduce_scatter_block_inter_sched_remote_reduce_local_scatterv(const vo
 
         /* reduce from right group to rank 0 */
         root = (rank == 0) ? MPI_ROOT : MPI_PROC_NULL;
-        mpi_errno = MPIR_Ireduce_sched_auto(sendbuf, tmp_buf, total_count,
-                                            datatype, op, root, comm_ptr, s);
+        mpi_errno = MPIR_Ireduce_inter_sched_auto(sendbuf, tmp_buf, total_count,
+                                                  datatype, op, root, comm_ptr, s);
         MPIR_ERR_CHECK(mpi_errno);
     }
     MPIR_SCHED_BARRIER(s);
@@ -87,8 +87,8 @@ int MPIR_Ireduce_scatter_block_inter_sched_remote_reduce_local_scatterv(const vo
 
     newcomm_ptr = comm_ptr->local_comm;
 
-    mpi_errno = MPIR_Iscatter_sched_auto(tmp_buf, recvcount, datatype,
-                                         recvbuf, recvcount, datatype, 0, newcomm_ptr, s);
+    mpi_errno = MPIR_Iscatter_intra_sched_auto(tmp_buf, recvcount, datatype,
+                                               recvbuf, recvcount, datatype, 0, newcomm_ptr, s);
     MPIR_ERR_CHECK(mpi_errno);
 
   fn_exit:
